refactor: deduplicate ApiError types into agentd-common crate
Move the ApiError enum and IntoResponse implementation from both
notify and orchestrator into the shared agentd-common error module.

Common ApiError variants:
- NotFound (404) — resource not found
- InvalidInput(String) (400) — bad request
- Conflict(String) (409) — resource conflict / invalid state
- Internal(anyhow::Error) (500) — internal server error

Migration:
- notify's Database(anyhow::Error) → Internal(anyhow::Error) (same 500)
- notify's NotFound(String) → NotFound (message dropped; 404 is sufficient)
- orchestrator's AgentNotRunning(String) → Conflict(String) (same 409)
- orchestrator's Internal(anyhow::Error) → unchanged
- Both crates' IntoResponse impls → single shared impl in common

Both crates re-export ApiError from their api.rs modules, so all
imports in scheduler/api.rs, cli, and tests continue working.

Changes:
- crates/common/src/error.rs: ApiError enum + IntoResponse + 4 tests
- crates/notify/Cargo.toml: add agentd-common dependency
- crates/notify/src/api.rs: replace local ApiError with re-export
- crates/orchestrator/Cargo.toml: add agentd-common dependency
- crates/orchestrator/src/api.rs: replace local ApiError with re-export,
  rename AgentNotRunning to Conflict

Closes #48
