+++ crates/orchestrator/src/api.rs
@@ -137,7 +137,7 @@ async fn send_message(
     // Verify agent exists and is running.
     let agent = state.manager.get_agent(&id).await?.ok_or(ApiError::NotFound)?;
     if agent.status != AgentStatus::Running {
-        return Err(ApiError::AgentNotRunning(format!(
+        return Err(ApiError::Conflict(format!(
             "Agent {} is not running (status: {})",
             id, agent.status
         )));
@@ -282,39 +282,5 @@ async fn list_agent_approvals(
 
 // -- Error handling --
 
-/// API error types for the orchestrator service.
-///
-/// # HTTP Status Mapping
-///
-/// - `NotFound` -> 404 Not Found
-/// - `InvalidInput` -> 400 Bad Request
-/// - `AgentNotRunning` -> 409 Conflict
-/// - `Internal` -> 500 Internal Server Error
-#[derive(Debug, thiserror::Error)]
-pub enum ApiError {
-    /// Resource not found (HTTP 404)
-    #[error("not found")]
-    NotFound,
-    /// Invalid input or request (HTTP 400)
-    #[error("invalid input: {0}")]
-    InvalidInput(String),
-    /// Agent is not in the expected state (HTTP 409)
-    #[error("agent not running: {0}")]
-    AgentNotRunning(String),
-    /// Internal server error (HTTP 500)
-    #[error(transparent)]
-    Internal(#[from] anyhow::Error),
-}
-
-impl IntoResponse for ApiError {
-    fn into_response(self) -> axum::response::Response {
-        let (status, message) = match &self {
-            ApiError::NotFound => (StatusCode::NOT_FOUND, self.to_string()),
-            ApiError::InvalidInput(_) => (StatusCode::BAD_REQUEST, self.to_string()),
-            ApiError::AgentNotRunning(_) => (StatusCode::CONFLICT, self.to_string()),
-            ApiError::Internal(e) => (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()),
-        };
-
-        (status, Json(serde_json::json!({ "error": message }))).into_response()
-    }
-}
+// Re-export shared ApiError from agentd-common.
+pub use agentd_common::error::ApiError;
